+++ bin/virtual_device.dart
@@ -8,22 +8,10 @@ import 'package:virtual_device/src/executable/stop_command.dart';
 import 'package:virtual_device/src/executable/wipe_command.dart';
 
 void main(List<String> args) async {
-  var runner = CommandRunner(
+  final runner = CommandRunner(
     'virtual_device',
     'Manage iOS simulators and Android emulators from the command line',
-  );
-  runner.argParser.addOption(
-    'platform',
-    allowed: ['ios', 'android'],
-    abbr: 'p',
-    defaultsTo: 'ios',
-    allowedHelp: {
-      'ios': 'Start an iOS device',
-      'android': 'Start an Android device'
-    },
-  );
-
-  runner = runner
+  )
     ..addCommand(CreateCommand())
     ..addCommand(DeleteCommand())
     ..addCommand(ListCommand())