+++ lib/src/executable/virtual_device_command.dart
@@ -5,5 +5,16 @@ abstract class VirtualDeviceCommand extends Command {
 
   bool get isIos => argResults['platform'] == 'ios';
 
-  VirtualDeviceCommand();
+  VirtualDeviceCommand() {
+    argParser.addOption(
+      'platform',
+      allowed: ['ios', 'android'],
+      abbr: 'p',
+      defaultsTo: 'ios',
+      allowedHelp: {
+        'ios': 'Start an iOS device',
+        'android': 'Start an Android device'
+      },
+    );
+  }
 }